+++ src/commands/execute-organization-changeset.ts
@@ -21,7 +21,6 @@ export class ExecuteChangeSetCommand extends BaseCliCommand<IExecuteChangeSetCom
     }
 
     public async performCommand(command: IExecuteChangeSetCommandArgs): Promise<void> {
-
         const changeSetName = command.changeSetName;
         const stateBucketName = await BaseCliCommand.GetStateBucketName(command.stateBucketName);
         const provider = new ChangeSetProvider(stateBucketName);